DriveSystem almost done & Light done.

+++ lib/Color/Color.cpp
@@ -27,35 +27,35 @@ along with Moti. If not, see <http://www.gnu.org/licenses/>.
  */
 
 Color::Color() {
-    _r = _g = _b = 0;
+	_r = _g = _b = 0;
 }
 
-Color::Color(uint8_t r, uint8_t g, uint8_t b) {
-    _r = r;
-    _g = g;
-    _b = b;
+Color::Color(int16_t r, int16_t g, int16_t b) {
+	_r = r;
+	_g = g;
+	_b = b;
 }
 
-void Color::getRGB(uint8_t* r, uint8_t* g, uint8_t* b) {
-    *r = _r;
-    *g = _g;
-    *b = _b;
+void Color::getRGB(int16_t* r, int16_t* g, int16_t* b) {
+	*r = _r;
+	*g = _g;
+	*b = _b;
 }
 
-uint8_t Color::getR(void) {
-    return _r;
+int16_t Color::getR(void) {
+	return _r;
 }
 
-uint8_t Color::getG(void) {
-    return _g;
+int16_t Color::getG(void) {
+	return _g;
 }
 
-uint8_t Color::getB(void) {
-    return _b;
+int16_t Color::getB(void) {
+	return _b;
 }
 
-void Color::setRGB(uint8_t r, uint8_t g, uint8_t b) {
-    _r = r;
-    _g = g;
-    _b = b;
+void Color::setRGB(int16_t r, int16_t g, int16_t b) {
+	_r = r;
+	_g = g;
+	_b = b;
 }
